std_svc: psci: introduce handler for system suspend
System Suspend needs to be executed on the last standing CPU and
psci_system_suspend() makes sure of that. The real problem with
the current code is that it acquires locks further into the CPU
suspend sequence even for the system suspend case. This seems
unnecessary and adds to the entry latency times.

This patch introduces a separate handler for the system suspend
case which does not acquire any locks and hence reduces the
entry time into the state.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>