Allow ARM common platforms to specify the base memory attribute
The memory attributes of the image sections are fixed on ARM common
platforms.
With this patch, you can define a platform specific memory type, which
enables platforms with images stored in non-cacheable memory. Both the
TCR and image translation table memory types are configured according to
the provided base memory type.

Change-Id: I43c132eaa1c03d37fef4dabab59bd1f7256f52b2